Added core folder and moved all the files to the core folder. Also added agent factory to handle the agent creation.
