+++ src/core/agents/__init__.py
@@ -1,0 +1,4 @@
+from src.core.agents.agent_factory import run_agent
+from src.core.agents.model_provider import MODEL, TOOL_ENABLED_MODEL
+
+__all__ = ["run_agent", "MODEL", "TOOL_ENABLED_MODEL"]
