+++ src/core/agents/custom_react_agent.py
@@ -1,18 +1,15 @@
-from langchain_openai import ChatOpenAI
-from langgraph.prebuilt import create_react_agent
-
 from src.config import Settings
-from src.services.agents.model_provider import MODEL
-from src.tools import TOOLS
 from src.utils.chat import print_event, get_ai_response
+from src.core.graphs.graph_builder import GRAPH
 from src.utils.logger import logger
 
 
-async def run_react_agent(user_input: str):
+async def run_custom_react_agent(thread_id: str, user_input: str):
     """
-    Run the standard react agent workflow.
+    Run the custom react agent workflow.
 
     Args:
+        thread_id (str): Unique identifier for the conversation thread.
         user_input (str): The input message from the user to be processed.
 
     Returns:
@@ -22,13 +19,11 @@ async def run_react_agent(user_input: str):
     prompt = settings.AGENT_CONFIG.get("prompt", "You are a helpful assistant.")
     logger.debug(f"System Prompt for React Agent: {prompt}")
 
-    # Create the React agent
-    graph = create_react_agent(model=MODEL, tools=TOOLS, prompt=prompt)
-
-    inputs = {"messages": [("user", user_input)]}
+    config = {"configurable": {"thread_id": thread_id}}
+    inputs = {"messages": [("user", user_input), ("system", prompt)]}
     events = []
 
-    async for event in graph.astream(inputs, stream_mode="values"):
+    async for event in GRAPH.astream(inputs, config=config, stream_mode="values"):
         print_event(event)
         events.append(event)
