+++ src/core/nodes/should_continue.py
