+++ src/core/nodes/tool_node.py
