+++ src/tools/__init__.py
@@ -1,4 +1,4 @@
-from src.utils.tool_registry import auto_register_tools
+from src.tools.tool_registry import auto_register_tools
 
 # TOOLS: List of BaseTool instances
 # TOOLS_BY_NAME: Dictionary of tool names and their corresponding BaseTool instances
