+++ src/tools/tool_registry.py
